% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_pages.R
\name{mot_get_page}
\alias{mot_get_page}
\title{Download a page of MOT data}
\usage{
mot_get_page(page, key, get_comments = TRUE, ...)
}
\arguments{
\item{page}{page number}

\item{key}{API Key}

\item{get_comments}{logical, should comments be downloaded, default TRUE}

\item{...}{extra variaibles}
}
\description{
Download one page of MOT data
}
