% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_reg.R
\name{mot_get_reg}
\alias{mot_get_reg}
\title{Download a registration of MOT data}
\usage{
mot_get_reg(registration, key, get_comments = TRUE, ...)
}
\arguments{
\item{registration}{registration number}

\item{key}{MOT api key}

\item{get_comments}{logical, should comments be downloaded, default TRUE}

\item{...}{extra variaibles}
}
\description{
Download one registration of MOT data
}
