% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_reg.R
\name{mot_get_regs}
\alias{mot_get_regs}
\title{Download multiple pages of MOT data}
\usage{
mot_get_regs(
  registrations,
  key = Sys.getenv("MOT_key"),
  ncores = 8L,
  get_comments = TRUE,
  ...
)
}
\arguments{
\item{registrations}{a range o page numberspage number}

\item{key}{API Key}

\item{ncores}{Now many cores to use}

\item{get_comments}{logical, should comments be downloaded, default TRUE}

\item{...}{extra variaibles passed to `mot_get_page()`}
}
\description{
Download one page of MOT data
}
\details{
This function wraps `mot_get_reg()` for bulk downloading of data, by default
   it is multicore optimised and will run with 8 cores as this is the upper 
   limit allowed by the API.
}
